requirements.txt
aiohttp>=3.9.1
asyncio>=3.4.3
python-dateutil>=2.8.2
python-dotenv>=1.0.0
